Fix for multiple windows
Breaks if there are multiple open windows and a different active tab is returned. 
Without the "tabs" permission, tab.url is not exposed and returns `undefined`. (would still be a wrong URL anyway)
I see there was a query object ready to fix this by restricting to the currentWindow, but it wasn't used.